requirements.txt
mcp
python-dotenv
pymysql
cryptography
DBUtils